Generalize osmWay
